requirements.txt
streamlit
opencv-python
plotly
numpy
pandas
pillow
scikit-learn
plotly